UPDATE: profile page with data from server
