+++ client/src/Components/PostDetail/index.tsx
@@ -2,16 +2,17 @@ import * as React from 'react';
 import { PageContainer } from '../../styles/page';
 import PostItem from '../PostItem';
 import MarkdownRenderer from '../MarkdownRenderer';
+import { Content } from './styles';
 
 function PostDetail(props: any) {
   const { post } = props;
-  console.log(post);
+
   return (
     <PageContainer width='80%'>
-      <PostItem content={post} />
-      <div style={{ height: '30rem', padding: '2%' }}>
+      <PostItem content={post} detail />
+      <Content>
         <MarkdownRenderer text={post.content} open />
-      </div>
+      </Content>
     </PageContainer>
   );
 }
