+++ client/src/Components/PostDetail/styles.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Content = styled.div`
+  height: 30rem;
+  padding: 2%;
+`;
