+++ client/src/Components/ProfileDetail/navTypes.ts
@@ -1,0 +1,32 @@
+export const navItem = [
+  {
+    menu: '내 프로필',
+    actionType: 'push',
+    url: '/myprofile',
+  },
+  {
+    menu: '관심있는 글',
+    actionType: 'alert',
+    message: '관심있는 글을 모아 보는 기능이 추가될 예정입니다.',
+  },
+  {
+    menu: '내가 쓴 댓글',
+    actionType: 'alert',
+    message: '내가 쓴 댓글을 한 번에 볼 수 있는 기능이 추가될 예정입니다.',
+  },
+  {
+    menu: '내가 쓴 글',
+    actionType: 'alert',
+    message: '내가 쓴 글을 한 번에 볼 수 있는 기능이 추가될 예정입니다',
+  },
+  {
+    menu: '로그아웃',
+    actionType: 'alert',
+    message: '로그아웃 기능 구현 예정',
+  },
+  {
+    menu: '탈퇴하기',
+    actionType: 'alert',
+    message: '탈퇴하기 기능 구현 예정',
+  },
+];
