+++ client/src/Components/ProfileDetail/styles.tsx
